opost: reshuffle settings.py
I converted settings.py by simply regenerating a fresh settings.py in
Django 4.x and porting all our settings to it.

This reorganizes the old settings module to match the approximate order
of Django 4.x's settings.py, so that the changes to settings.py in the
upgrade will be more obvious.